feat: Define the virtualization layer of local k8s
Vagrant is used to provision the VMs, which are configured via a YAML
file `config.yml`.

The configuration is based on what the Kubernetes documentation sets as
minimum requirements.

Also, a simple README is added to explain the requirements and the basic
usage.

The README steps will probably change with later commits.