feat: Create the necessary network and QEMU setup for VMs
QEMU is used to bootstrap each VM that the local K8s cluster will use.

In order to setup the networking, `dnsmasq` is used along with a
workaround solution to disable the default DNS & DHCP server on macOS.

Placeholder `cloud-init` files are added for each type of VM for future
use.

A README will be added for detailed instructions.