feat: Setup SSH for the VMs
`cloud-init` scripts are updated to contain the public SSH key pair
generated for the VMs, called `vm-ssh-key[.pub]`.

A new file called `setup-ssh` is implemented to automate the SSH
configuration, meaning that the stale entries in `known_hosts` are
removed whenever this script is executed after the VM is booted up.

Also, the `helpers` script is renamed to `definitions` to better suit
the usage of the script.

Since SSH setup needs inet configurations as well, these definitions are
moved under the `definitions` script.