feat: Bootstrap security via generating and deploying TLS certs
Before starting to install and use K8s components, the security aspect
of the cluster is improved by automating the creation and deployment of
TLS certificates that encrypts the communication between the components.

The certificate generation is handled in `./certs/generate` script.
The certificate deployment is handled in `./certs/deploy` script.

---

The details of TLS certificates and the usages of these scripts will be
explained in README later.