feat: Add control-plane setup script to install and configure binaries
The following binaries are installed and configured:

- `etcd`
- `kube-apiserver`
- `kube-controller-manager`
- `kube-scheduler`

The key part in most of the script is the TLS configuration. Instead of
using separate CA and cert/key pairs for each binary, one CA and
cert/key pair is used instead.

Basically, the control plane only communication is secured by one CA
and one cert/key pair.

This lowers the security aspects of the cluster but for local usage, as
long as the consequences are known I believe it is fine to do so.