feat: Configure LB for control-plane nodes through gateway
Gateway machine is set up as a load balancer between the control nodes
through the virtual IP `192.168.200.11` and a binary called
`ldirectord`.

This IP was left outside the DHCP range intentionally, because it is
owned by 4 different machines (gateway + control{0..2}). Hence, it
becomes a virtual IP.

In order to make gateway the only visible owner of this virtual IP, the
ARP request & response messages of control nodes are updated via
`cloud-init/network-config.control{0..2}`.

As a temporary solution, the DNS lookup for `192.168.200.11` (kube-api)
was disabled, because the default DNS server of the nodes cannot resolve
this name as of now (127.0.0.1:53 instead of 192.168.200.1:53).

This will be fixed in a later commit, and afterwards `kube-api` will be
used instead of hardcoded IP's in kubeconfigs and other places.