feat: Add the setup script for the cluster
Along with the remove script, a setup script is also added to the
repository to create a running cluster with only one command.

This script aggregates every other script in this repository in a
specific order that allows the user to have a running Kubernetes
cluster.

The existing scripts are updated along the way to adjust the usage and
fix runtime errors.

Also, certs directory is renamed to auth since it contains more than TLS
certificates (e.g. kubeconfigs, encryption-key).