read: Simplify the README's
Since the only noticable difference between the guide and this
implementation is the networking part, the other README's that touch on
the rest of the cluster setup are removed.

There is no need to duplicate the information presented in the guide as
the guide itself is a REALLY nice read.